Updating tests for stl pov converter now that it converts natively to PHP
